Link datasets w/ extension into working directory
In the context of #3145, it would be nice if tool wrapper authors could
rely on the datatype extension for their tools. Many tools require a
specific filename extension. Currently, tool authors need to symlink the
input files with the correct extension in the tool wrapper.

With this commit galaxy will symlink input files into the current
working directory as input_<dataset_id>.<datatype_extension>, if an
extension is defined, the datatype is not composite and the tool does
not make use of parallelism.

I believe this makes writing wrappers easier for tool authors, and
should also prevent resorting to cheetah if a tool accepts different
input formats, where each format has a specific extension (e.g fastq,
fastq.gz. fastq.bz2 etc).