Add a cache for known broken tools
Attempting to load broken tools (malformed XML, no TS repository available) can
take a considerable amount of time when reloading the toolbox. If we are unable
to reload the tool because we can't create a Tool instance we add it as a known
broken tool to the tool cache and don't attempt to reinitialize the tool unless
the tool has been updated on disk.

Previously we would only cache a tool if it had an associated tool shed repository.
We also now cache a tool whenever it was successfully loaded. If a tool shed
tool does not have an associated tool shed repository record, it will still
fail to load (as previously), but when reloading the toolbox we don't need to
re-parse the tool xml.  This should help on instances where many broken tools
are installed.