Allow drag and drop datasets from collections to the tool form.
There are some shortcomings here - the original name will still be used in the tool since we are treating the dataset as a stand-alone dataset and not an element in the collection for instance. This is an issue in the display of the selected dataset and in tools that may want to leverage element_identifier. Still - this prevents the user from having to find the right dataset and unhide it to use it in the tool form - good enough for a step forward?