Fix map over with structured_like collection outputs
The problem seems to be that self.mapping_params.param_template has not been
expanded entirely, so in some instances there's no collection to infer the
collection_type from, while self.example_params is completely expanded.

An example param_template may look like this:

{'on_string': 'collection 3', 'input': {'values': [{'src': u'hdca', 'tags': [], 'hid': 3, 'id': u'3509a83a05cae1d5', 'name': u'c'}], 'batch': False}, 'tool': <galaxy.tools.Tool object at 0x10e422490>, 'how|how_filter': u'remove_if_absent', 'how|filter_source': <galaxy.model.HistoryDatasetCollectionAssociation object at 0x11141f510>}
If we have an output that is `structured_like="input"` we will fail here.